refactor(metadata): split repository ops
