+++ src/metadata/meson.build
@@ -12,8 +12,15 @@ yams_metadata = static_library('yams_metadata',
     'migration.cpp',
     'metadata_repository.cpp',
     'repository/content_batch_ops.cpp',
+    'repository/content_ops.cpp',
+    'repository/document_crud_ops.cpp',
     'repository/document_query_filters.cpp',
+    'repository/document_status_ops.cpp',
+    'repository/history_ops.cpp',
+    'repository/metadata_kv_ops.cpp',
     'repository/metadata_value_count_ops.cpp',
+    'repository/relationship_ops.cpp',
+    'repository/path_tree_ops.cpp',
     'repository/search_ops.cpp',
     'repository/search_query_helpers.cpp',
     'query_helpers.cpp',
