+++ src/metadata/repository/document_status_ops.cpp
@@ -1,0 +1,788 @@
+// Copyright (c) 2025 YAMS Contributors
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include <algorithm>
+#include <atomic>
+#include <chrono>
+#include <cstdint>
+#include <random>
+#include <string>
+#include <thread>
+#include <vector>
+
+#include <yams/core/atomic_utils.h>
+#include <yams/metadata/metadata_repository.h>
+#include <yams/storage/sqlite_retry.h>
+
+#include "crud_ops.hpp"
+#include "transaction_helpers.hpp"
+
+namespace yams::metadata {
+
+using repository::beginTransactionWithRetry;
+
+Result<int64_t> MetadataRepository::getEmbeddedDocumentCount() {
+    return executeReadQuery<int64_t>([&](Database& db) -> Result<int64_t> {
+        auto stmtResult =
+            db.prepare("SELECT COUNT(*) FROM document_embeddings_status WHERE has_embedding = 1");
+        if (!stmtResult)
+            return stmtResult.error();
+        auto& stmt = stmtResult.value();
+        auto stepResult = stmt.step();
+        if (!stepResult)
+            return stepResult.error();
+        return stepResult.value() ? stmt.getInt64(0) : int64_t{0};
+    });
+}
+
+Result<int64_t> MetadataRepository::getDocumentCountByExtractionStatus(ExtractionStatus status) {
+    return executeReadQuery<int64_t>([&](Database& db) -> Result<int64_t> {
+        repository::CrudOps<DocumentInfo> ops;
+        return ops.count(db, "extraction_status = ?", ExtractionStatusUtils::toString(status));
+    });
+}
+
+Result<void> MetadataRepository::updateDocumentEmbeddingStatus(int64_t documentId,
+                                                               bool hasEmbedding,
+                                                               const std::string& modelId) {
+    return executeQuery<void>([&](Database& db) -> Result<void> {
+        // Check current embedding status to track changes
+        bool hadEmbedding = false;
+        {
+            auto checkStmt = db.prepare("SELECT COALESCE(has_embedding, 0) FROM "
+                                        "document_embeddings_status WHERE document_id = ?");
+            if (checkStmt) {
+                auto& stmt = checkStmt.value();
+                stmt.bind(1, documentId);
+                if (auto stepRes = stmt.step(); stepRes && stepRes.value()) {
+                    hadEmbedding = stmt.getInt(0) != 0;
+                }
+            }
+        }
+
+        // Ensure the document exists
+        auto docCheckStmt = db.prepare("SELECT 1 FROM documents WHERE id = ?");
+        if (!docCheckStmt)
+            return docCheckStmt.error();
+
+        auto& stmt = docCheckStmt.value();
+        if (auto r = stmt.bind(1, documentId); !r)
+            return r.error();
+
+        auto stepResult = stmt.step();
+        if (!stepResult)
+            return stepResult.error();
+        if (!stepResult.value()) {
+            return Error{ErrorCode::NotFound, "Document not found"};
+        }
+
+        // Ensure model_id exists in vector_models (FK constraint)
+        if (!modelId.empty()) {
+            auto ensureModelStmt = db.prepare(R"(
+                INSERT OR IGNORE INTO vector_models (model_id, model_name, embedding_dim)
+                VALUES (?, ?, 0)
+            )");
+            if (ensureModelStmt) {
+                auto& mstmt = ensureModelStmt.value();
+                mstmt.bind(1, modelId);
+                mstmt.bind(2, modelId); // Use model_id as name if not registered
+                (void)mstmt.execute();
+            }
+        }
+
+        // Insert or update the embedding status
+        auto updateStmt = db.prepare(R"(
+                INSERT INTO document_embeddings_status (document_id, has_embedding, model_id, updated_at)
+                VALUES (?, ?, ?, unixepoch())
+                ON CONFLICT(document_id) DO UPDATE SET
+                    has_embedding = excluded.has_embedding,
+                    model_id = excluded.model_id,
+                    updated_at = excluded.updated_at
+            )");
+        if (!updateStmt)
+            return updateStmt.error();
+
+        auto& ustmt = updateStmt.value();
+        if (auto r = ustmt.bind(1, documentId); !r)
+            return r.error();
+        if (auto r = ustmt.bind(2, hasEmbedding ? 1 : 0); !r)
+            return r.error();
+        if (auto r = ustmt.bind(3, modelId.empty() ? nullptr : modelId.c_str()); !r)
+            return r.error();
+
+        auto execResult = ustmt.execute();
+        if (!execResult)
+            return execResult.error();
+
+        // Note: cachedIndexedCount_ now tracks actual FTS5 rows, not extraction success or
+        // embedding status. Embedding status is tracked separately via
+        // VectorDatabase::getVectorCount()
+
+        // Update component-owned embedded-doc counter only on state transition.
+        if (!hadEmbedding && hasEmbedding) {
+            cachedEmbeddedCount_.fetch_add(1, std::memory_order_relaxed);
+            signalCorpusStatsStale();
+        } else if (hadEmbedding && !hasEmbedding) {
+            core::saturating_sub(cachedEmbeddedCount_, uint64_t{1});
+            signalCorpusStatsStale();
+        }
+
+        return Result<void>();
+    });
+}
+
+Result<void> MetadataRepository::updateDocumentEmbeddingStatusByHash(const std::string& hash,
+                                                                     bool hasEmbedding,
+                                                                     const std::string& modelId) {
+    return executeQuery<void>([&](Database& db) -> Result<void> {
+        // Get document ID and previous embedding status in one query.
+        auto getIdStmt = db.prepare(R"(
+            SELECT d.id, COALESCE(des.has_embedding, 0)
+            FROM documents d
+            LEFT JOIN document_embeddings_status des ON d.id = des.document_id
+            WHERE d.sha256_hash = ?
+        )");
+        if (!getIdStmt)
+            return getIdStmt.error();
+
+        auto& stmt = getIdStmt.value();
+        if (auto r = stmt.bind(1, hash); !r)
+            return r.error();
+
+        auto stepResult = stmt.step();
+        if (!stepResult)
+            return stepResult.error();
+        if (!stepResult.value()) {
+            return Error{ErrorCode::NotFound, "Document with hash not found"};
+        }
+
+        int64_t documentId = stmt.getInt64(0);
+        bool hadEmbedding = stmt.getInt(1) != 0;
+
+        // Ensure model_id exists in vector_models (FK constraint)
+        if (!modelId.empty()) {
+            auto ensureModelStmt = db.prepare(R"(
+                INSERT OR IGNORE INTO vector_models (model_id, model_name, embedding_dim)
+                VALUES (?, ?, 0)
+            )");
+            if (ensureModelStmt) {
+                auto& mstmt = ensureModelStmt.value();
+                mstmt.bind(1, modelId);
+                mstmt.bind(2, modelId); // Use model_id as name if not registered
+                (void)mstmt.execute();
+            }
+        }
+
+        // Insert or update the embedding status
+        auto updateStmt = db.prepare(R"(
+                INSERT INTO document_embeddings_status (document_id, has_embedding, model_id, updated_at)
+                VALUES (?, ?, ?, unixepoch())
+                ON CONFLICT(document_id) DO UPDATE SET
+                    has_embedding = excluded.has_embedding,
+                    model_id = excluded.model_id,
+                    updated_at = excluded.updated_at
+            )");
+        if (!updateStmt)
+            return updateStmt.error();
+
+        auto& ustmt = updateStmt.value();
+        if (auto r = ustmt.bind(1, documentId); !r)
+            return r.error();
+        if (auto r = ustmt.bind(2, hasEmbedding ? 1 : 0); !r)
+            return r.error();
+        if (auto r = ustmt.bind(3, modelId.empty() ? nullptr : modelId.c_str()); !r)
+            return r.error();
+
+        auto execResult = ustmt.execute();
+        if (!execResult)
+            return execResult.error();
+
+        // Update component-owned embedded-doc counter only on state transition.
+        if (!hadEmbedding && hasEmbedding) {
+            cachedEmbeddedCount_.fetch_add(1, std::memory_order_relaxed);
+            signalCorpusStatsStale();
+        } else if (hadEmbedding && !hasEmbedding) {
+            core::saturating_sub(cachedEmbeddedCount_, uint64_t{1});
+            signalCorpusStatsStale();
+        }
+
+        return Result<void>();
+    });
+}
+
+Result<void> MetadataRepository::batchUpdateDocumentEmbeddingStatusByHashes(
+    const std::vector<std::string>& hashes, bool hasEmbedding, const std::string& modelId) {
+    if (hashes.empty())
+        return Result<void>();
+
+    constexpr int kMaxRetries = 7; // Increased for heavy concurrent load
+    constexpr int kBaseDelayMs = 50;
+
+    // Thread-local RNG for jitter to avoid thundering herd
+    thread_local std::mt19937 rng(std::random_device{}());
+
+    for (int attempt = 0; attempt < kMaxRetries; ++attempt) {
+        auto result = executeQuery<void>([&](Database& db) -> Result<void> {
+#if YAMS_LIBSQL_BACKEND
+            auto beginResult = db.execute("BEGIN");
+#else
+            auto beginResult = db.execute("BEGIN IMMEDIATE");
+#endif
+            if (!beginResult)
+                return beginResult.error();
+
+            // Ensure model_id exists in vector_models (FK constraint) - once per batch
+            if (!modelId.empty()) {
+                auto ensureModelStmt = db.prepare(R"(
+                    INSERT OR IGNORE INTO vector_models (model_id, model_name, embedding_dim)
+                    VALUES (?, ?, 0)
+                )");
+                if (ensureModelStmt) {
+                    auto& mstmt = ensureModelStmt.value();
+                    mstmt.bind(1, modelId);
+                    mstmt.bind(2, modelId); // Use model_id as name if not registered
+                    (void)mstmt.execute();
+                }
+            }
+
+            auto lookupStmt = db.prepare(R"(
+                SELECT d.id, COALESCE(des.has_embedding, 0)
+                FROM documents d
+                LEFT JOIN document_embeddings_status des ON d.id = des.document_id
+                WHERE d.sha256_hash = ?
+            )");
+            if (!lookupStmt) {
+                db.execute("ROLLBACK");
+                return lookupStmt.error();
+            }
+
+            auto updateStmt = db.prepare(R"(
+                INSERT INTO document_embeddings_status (document_id, has_embedding, model_id, updated_at)
+                VALUES (?, ?, ?, unixepoch())
+                ON CONFLICT(document_id) DO UPDATE SET
+                    has_embedding = excluded.has_embedding,
+                    model_id = excluded.model_id,
+                    updated_at = excluded.updated_at
+            )");
+            if (!updateStmt) {
+                db.execute("ROLLBACK");
+                return updateStmt.error();
+            }
+
+            auto& lstmt = lookupStmt.value();
+            auto& ustmt = updateStmt.value();
+            int64_t embeddedDelta = 0;
+
+            for (const auto& hash : hashes) {
+                lstmt.reset();
+                if (auto r = lstmt.bind(1, hash); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+
+                auto stepResult = lstmt.step();
+                if (!stepResult) {
+                    db.execute("ROLLBACK");
+                    return stepResult.error();
+                }
+                if (!stepResult.value())
+                    continue;
+
+                int64_t documentId = lstmt.getInt64(0);
+                bool hadEmbedding = lstmt.getInt(1) != 0;
+
+                ustmt.reset();
+                if (auto r = ustmt.bind(1, documentId); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+                if (auto r = ustmt.bind(2, hasEmbedding ? 1 : 0); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+                if (auto r = ustmt.bind(3, modelId.empty() ? nullptr : modelId.c_str()); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+
+                auto execResult = ustmt.execute();
+                if (!execResult) {
+                    db.execute("ROLLBACK");
+                    return execResult.error();
+                }
+
+                if (!hadEmbedding && hasEmbedding) {
+                    ++embeddedDelta;
+                } else if (hadEmbedding && !hasEmbedding) {
+                    --embeddedDelta;
+                }
+            }
+
+            auto commitResult = db.execute("COMMIT");
+            if (!commitResult)
+                return commitResult.error();
+
+            if (embeddedDelta > 0) {
+                cachedEmbeddedCount_.fetch_add(static_cast<uint64_t>(embeddedDelta),
+                                               std::memory_order_relaxed);
+            } else if (embeddedDelta < 0) {
+                core::saturating_sub(cachedEmbeddedCount_, static_cast<uint64_t>(-embeddedDelta));
+            }
+
+            signalCorpusStatsStale();
+            return Result<void>();
+        });
+
+        if (result)
+            return result;
+
+        if (!storage::sqlite_retry::isBusyOrLockedMessage(result.error().message))
+            return result;
+
+        // Exponential backoff with jitter (±25%) to prevent thundering herd
+        int baseDelayMs = kBaseDelayMs * (1 << attempt);
+        int jitter = static_cast<int>(baseDelayMs * 0.25);
+        std::uniform_int_distribution<int> dist(-jitter, jitter);
+        int delayMs = baseDelayMs + dist(rng);
+        std::this_thread::sleep_for(std::chrono::milliseconds(delayMs));
+    }
+
+    daemon::TuneAdvisor::reportDbLockError();
+    return Error{ErrorCode::DatabaseError,
+                 "batchUpdateDocumentEmbeddingStatusByHashes: max retries exceeded"};
+}
+
+Result<void> MetadataRepository::reconcileDocumentEmbeddingStatusByHashes(
+    const std::vector<std::string>& embeddedHashes, const std::string& modelId) {
+    constexpr int kMaxRetries = 5;
+    constexpr int kBaseDelayMs = 50;
+
+    std::vector<std::string> uniqueHashes = embeddedHashes;
+    std::sort(uniqueHashes.begin(), uniqueHashes.end());
+    uniqueHashes.erase(std::unique(uniqueHashes.begin(), uniqueHashes.end()), uniqueHashes.end());
+
+    thread_local std::mt19937 rng(std::random_device{}());
+    std::size_t reconciledEmbeddedDocs = 0;
+
+    for (int attempt = 0; attempt < kMaxRetries; ++attempt) {
+        auto result = executeQuery<void>([&](Database& db) -> Result<void> {
+#if YAMS_LIBSQL_BACKEND
+            auto beginResult = db.execute("BEGIN");
+#else
+            auto beginResult = db.execute("BEGIN IMMEDIATE");
+#endif
+            if (!beginResult)
+                return beginResult.error();
+
+            auto rollback = [&db](const Result<void>& err) -> Result<void> {
+                db.execute("ROLLBACK");
+                return err;
+            };
+
+            if (auto createTemp =
+                    db.execute("CREATE TEMP TABLE IF NOT EXISTS temp_embedding_reconcile_hashes ("
+                               "sha256_hash TEXT PRIMARY KEY)");
+                !createTemp) {
+                return rollback(createTemp.error());
+            }
+            if (auto clearTemp = db.execute("DELETE FROM temp_embedding_reconcile_hashes");
+                !clearTemp) {
+                return rollback(clearTemp.error());
+            }
+
+            if (!modelId.empty()) {
+                auto ensureModelStmt = db.prepare(R"(
+                    INSERT OR IGNORE INTO vector_models (model_id, model_name, embedding_dim)
+                    VALUES (?, ?, 0)
+                )");
+                if (!ensureModelStmt) {
+                    return rollback(ensureModelStmt.error());
+                }
+                auto& mstmt = ensureModelStmt.value();
+                if (auto r = mstmt.bind(1, modelId); !r) {
+                    return rollback(r.error());
+                }
+                if (auto r = mstmt.bind(2, modelId); !r) {
+                    return rollback(r.error());
+                }
+                if (auto exec = mstmt.execute(); !exec) {
+                    return rollback(exec.error());
+                }
+            }
+
+            if (!uniqueHashes.empty()) {
+                auto insertHashStmt = db.prepare(
+                    "INSERT OR IGNORE INTO temp_embedding_reconcile_hashes (sha256_hash) "
+                    "VALUES (?)");
+                if (!insertHashStmt) {
+                    return rollback(insertHashStmt.error());
+                }
+                auto& hstmt = insertHashStmt.value();
+                for (const auto& hash : uniqueHashes) {
+                    hstmt.reset();
+                    if (auto r = hstmt.bind(1, hash); !r) {
+                        return rollback(r.error());
+                    }
+                    if (auto exec = hstmt.execute(); !exec) {
+                        return rollback(exec.error());
+                    }
+                }
+            }
+
+            auto countStmt = db.prepare(R"(
+                SELECT COUNT(*)
+                FROM documents d
+                JOIN temp_embedding_reconcile_hashes th ON th.sha256_hash = d.sha256_hash
+            )");
+            if (!countStmt) {
+                return rollback(countStmt.error());
+            }
+            if (auto step = countStmt.value().step(); !step) {
+                return rollback(step.error());
+            } else if (step.value()) {
+                reconciledEmbeddedDocs = static_cast<std::size_t>(countStmt.value().getInt64(0));
+            }
+
+            auto reconcileStmt = db.prepare(R"(
+                INSERT INTO document_embeddings_status (
+                    document_id, has_embedding, model_id, chunk_count, updated_at
+                )
+                SELECT d.id,
+                       CASE WHEN th.sha256_hash IS NOT NULL THEN 1 ELSE 0 END,
+                       CASE WHEN th.sha256_hash IS NOT NULL THEN ? ELSE NULL END,
+                       CASE WHEN th.sha256_hash IS NOT NULL THEN 1 ELSE 0 END,
+                       unixepoch()
+                FROM documents d
+                LEFT JOIN temp_embedding_reconcile_hashes th ON th.sha256_hash = d.sha256_hash
+                ON CONFLICT(document_id) DO UPDATE SET
+                    has_embedding = excluded.has_embedding,
+                    model_id = CASE
+                        WHEN excluded.has_embedding = 0 THEN NULL
+                        WHEN excluded.model_id IS NOT NULL THEN excluded.model_id
+                        ELSE document_embeddings_status.model_id
+                    END,
+                    chunk_count = CASE
+                        WHEN excluded.has_embedding = 0 THEN 0
+                        WHEN document_embeddings_status.chunk_count > 0
+                            THEN document_embeddings_status.chunk_count
+                        ELSE excluded.chunk_count
+                    END,
+                    updated_at = excluded.updated_at
+            )");
+            if (!reconcileStmt) {
+                return rollback(reconcileStmt.error());
+            }
+            auto& rstmt = reconcileStmt.value();
+            if (auto r = rstmt.bind(1, modelId.empty() ? nullptr : modelId.c_str()); !r) {
+                return rollback(r.error());
+            }
+            if (auto exec = rstmt.execute(); !exec) {
+                return rollback(exec.error());
+            }
+
+            if (auto clearTemp = db.execute("DELETE FROM temp_embedding_reconcile_hashes");
+                !clearTemp) {
+                return rollback(clearTemp.error());
+            }
+
+            auto commitResult = db.execute("COMMIT");
+            if (!commitResult) {
+                db.execute("ROLLBACK");
+                return commitResult.error();
+            }
+            return Result<void>();
+        });
+
+        if (result) {
+            cachedEmbeddedCount_.store(static_cast<uint64_t>(reconciledEmbeddedDocs),
+                                       std::memory_order_relaxed);
+            signalCorpusStatsStale();
+            return result;
+        }
+
+        if (!storage::sqlite_retry::isBusyOrLockedMessage(result.error().message)) {
+            return result;
+        }
+
+        int baseDelayMs = kBaseDelayMs * (1 << attempt);
+        int jitter = static_cast<int>(baseDelayMs * 0.25);
+        std::uniform_int_distribution<int> dist(-jitter, jitter);
+        int delayMs = baseDelayMs + dist(rng);
+        std::this_thread::sleep_for(std::chrono::milliseconds(delayMs));
+    }
+
+    daemon::TuneAdvisor::reportDbLockError();
+    return Error{ErrorCode::DatabaseError,
+                 "reconcileDocumentEmbeddingStatusByHashes: max retries exceeded"};
+}
+
+Result<bool> MetadataRepository::hasDocumentEmbeddingByHash(const std::string& hash) {
+    return executeReadQuery<bool>([&](Database& db) -> Result<bool> {
+        auto stmtResult = db.prepare(R"(
+            SELECT COALESCE(des.has_embedding, 0)
+            FROM documents d
+            LEFT JOIN document_embeddings_status des ON d.id = des.document_id
+            WHERE d.sha256_hash = ?
+        )");
+        if (!stmtResult)
+            return stmtResult.error();
+
+        auto& stmt = stmtResult.value();
+        if (auto r = stmt.bind(1, hash); !r)
+            return r.error();
+
+        auto stepResult = stmt.step();
+        if (!stepResult)
+            return stepResult.error();
+
+        if (!stepResult.value()) {
+            // Document not found - treat as no embedding
+            return false;
+        }
+
+        return stmt.getInt(0) != 0;
+    });
+}
+
+Result<void> MetadataRepository::updateDocumentExtractionStatus(int64_t documentId,
+                                                                bool contentExtracted,
+                                                                ExtractionStatus status,
+                                                                const std::string& error) {
+    return batchUpdateDocumentExtractionStatuses(
+        {ExtractionStatusUpdate{documentId, contentExtracted, status, error}});
+}
+
+Result<void> MetadataRepository::batchUpdateDocumentExtractionStatuses(
+    const std::vector<ExtractionStatusUpdate>& updates) {
+    if (updates.empty())
+        return Result<void>();
+
+    struct ExtractionDelta {
+        std::uint64_t newlyExtracted = 0;
+        std::uint64_t newlyUnextracted = 0;
+        std::uint64_t rowsUpdated = 0;
+    };
+
+    auto result = executeQuery<ExtractionDelta>([&](Database& db) -> Result<ExtractionDelta> {
+        auto beginResult = beginTransactionWithRetry(db);
+        if (!beginResult) {
+            return beginResult.error();
+        }
+
+        auto checkStmt = db.prepareCached(R"(
+            SELECT COALESCE(content_extracted, 0)
+            FROM documents WHERE id = ?
+        )");
+        if (!checkStmt) {
+            db.execute("ROLLBACK");
+            return checkStmt.error();
+        }
+
+        auto updateStmt = db.prepareCached(R"(
+            UPDATE documents
+            SET content_extracted = ?, extraction_status = ?, extraction_error = ?
+            WHERE id = ?
+        )");
+        if (!updateStmt) {
+            db.execute("ROLLBACK");
+            return updateStmt.error();
+        }
+
+        auto& check = *checkStmt.value();
+        auto& update = *updateStmt.value();
+        ExtractionDelta delta;
+
+        for (const auto& item : updates) {
+            if (item.documentId < 0) {
+                continue;
+            }
+
+            if (auto r = check.reset(); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+            if (auto r = check.bind(1, item.documentId); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+
+            auto stepResult = check.step();
+            if (!stepResult) {
+                db.execute("ROLLBACK");
+                return stepResult.error();
+            }
+            if (!stepResult.value()) {
+                continue;
+            }
+
+            const bool extractedBefore = check.getInt(0) != 0;
+
+            if (auto r = update.reset(); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+            if (auto r = update.bind(1, item.contentExtracted ? 1 : 0); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+            if (auto r = update.bind(2, ExtractionStatusUtils::toString(item.status)); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+            if (auto r = update.bind(3, item.error.empty() ? nullptr : item.error.c_str()); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+            if (auto r = update.bind(4, item.documentId); !r) {
+                db.execute("ROLLBACK");
+                return r.error();
+            }
+
+            auto execResult = update.execute();
+            if (!execResult) {
+                db.execute("ROLLBACK");
+                return execResult.error();
+            }
+
+            if (db.changes() > 0) {
+                ++delta.rowsUpdated;
+                if (!extractedBefore && item.contentExtracted) {
+                    ++delta.newlyExtracted;
+                } else if (extractedBefore && !item.contentExtracted) {
+                    ++delta.newlyUnextracted;
+                }
+            }
+        }
+
+        auto commitResult = db.execute("COMMIT");
+        if (!commitResult) {
+            db.execute("ROLLBACK");
+            return commitResult.error();
+        }
+
+        return delta;
+    });
+
+    if (!result) {
+        return result.error();
+    }
+
+    const auto& delta = result.value();
+    if (delta.newlyExtracted > 0) {
+        cachedExtractedCount_.fetch_add(delta.newlyExtracted, std::memory_order_relaxed);
+    }
+    if (delta.newlyUnextracted > 0) {
+        core::saturating_sub(cachedExtractedCount_, delta.newlyUnextracted);
+    }
+    if (delta.rowsUpdated > 0) {
+        signalCorpusStatsStale();
+    }
+
+    return Result<void>();
+}
+
+Result<void> MetadataRepository::updateDocumentRepairStatus(const std::string& hash,
+                                                            RepairStatus status) {
+    return executeQuery<void>([&](Database& db) -> Result<void> {
+        auto updateStmt = db.prepare(R"(
+            UPDATE documents
+            SET repair_status = ?, repair_attempted_at = unixepoch(), repair_attempts = repair_attempts + 1
+            WHERE sha256_hash = ?
+        )");
+        if (!updateStmt)
+            return updateStmt.error();
+
+        auto& stmt = updateStmt.value();
+        if (auto r = stmt.bind(1, RepairStatusUtils::toString(status)); !r)
+            return r.error();
+        if (auto r = stmt.bind(2, hash); !r)
+            return r.error();
+
+        auto execResult = stmt.execute();
+        if (!execResult)
+            return execResult.error();
+
+        return Result<void>();
+    });
+}
+
+Result<void>
+MetadataRepository::batchUpdateDocumentRepairStatuses(const std::vector<std::string>& hashes,
+                                                      RepairStatus status) {
+    if (hashes.empty())
+        return Result<void>();
+
+    constexpr int kMaxRetries = 7; // Increased for heavy concurrent load
+    constexpr int kBaseDelayMs = 50;
+
+    // Thread-local RNG for jitter to avoid thundering herd
+    thread_local std::mt19937 rng(std::random_device{}());
+
+    for (int attempt = 0; attempt < kMaxRetries; ++attempt) {
+        auto result = executeQuery<void>([&](Database& db) -> Result<void> {
+#if YAMS_LIBSQL_BACKEND
+            auto beginResult = db.execute("BEGIN");
+#else
+            auto beginResult = db.execute("BEGIN IMMEDIATE");
+#endif
+            if (!beginResult)
+                return beginResult.error();
+
+            auto updateStmt = db.prepare(R"(
+                UPDATE documents
+                SET repair_status = ?, repair_attempted_at = unixepoch(), repair_attempts = repair_attempts + 1
+                WHERE sha256_hash = ?
+            )");
+            if (!updateStmt) {
+                db.execute("ROLLBACK");
+                return updateStmt.error();
+            }
+
+            auto& stmt = updateStmt.value();
+            std::string statusStr = RepairStatusUtils::toString(status);
+
+            for (const auto& hash : hashes) {
+                stmt.reset();
+                if (auto r = stmt.bind(1, statusStr.c_str()); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+                if (auto r = stmt.bind(2, hash); !r) {
+                    db.execute("ROLLBACK");
+                    return r.error();
+                }
+
+                auto execResult = stmt.execute();
+                if (!execResult) {
+                    db.execute("ROLLBACK");
+                    return execResult.error();
+                }
+            }
+
+            auto commitResult = db.execute("COMMIT");
+            if (!commitResult)
+                return commitResult.error();
+
+            return Result<void>();
+        });
+
+        if (result)
+            return result;
+
+        if (!storage::sqlite_retry::isBusyOrLockedMessage(result.error().message))
+            return result;
+
+        // Exponential backoff with jitter (±25%) to prevent thundering herd
+        int baseDelayMs = kBaseDelayMs * (1 << attempt);
+        int jitter = static_cast<int>(baseDelayMs * 0.25);
+        std::uniform_int_distribution<int> dist(-jitter, jitter);
+        int delayMs = baseDelayMs + dist(rng);
+        std::this_thread::sleep_for(std::chrono::milliseconds(delayMs));
+    }
+
+    daemon::TuneAdvisor::reportDbLockError();
+    return Error{ErrorCode::DatabaseError,
+                 "batchUpdateDocumentRepairStatuses: max retries exceeded"};
+}
+
+} // namespace yams::metadata
